requirements.txt
accelerate
git+https://github.com/huggingface/diffusers.git
gradio==4.21.0
gradio_client==0.12.0
tqdm==4.66.5
transformers==4.43.3
numpy==1.26.4
peft==0.15.0
huggingface-hub
spaces
protobuf
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
bitsandbytes==0.44.1
sentencepiece==0.2.0
deepspeed
pandas
wandb
prodigyopt
supervision
pytest
einops
timm
uvicorn
fastapi
pydantic
opencv-python
Pillow
